+++ site/config.js
@@ -1,0 +1,7 @@
+window.DOCS_FORGE_CONFIG = Object.freeze({
+  policyVersion: "2026-07-29",
+  formEndpoint: "",
+  consentEndpoint: "",
+  posthogKey: "",
+  posthogHost: "https://us.i.posthog.com"
+});